slice: extract tile create_descriptor into a standalone builder
Move SliceTileProgramFactory::create_descriptor's body into a free function
build_slice_tile_descriptor(). The factory now delegates to it, and the pybind
(Python name unchanged) calls the free function directly.

This decouples the descriptor builder from the factory so the bind survives a
future Metal 2.0 port: a factory may not expose both create_descriptor and
create_program_spec (operation_concepts.hpp all_factories_valid requires exactly
one concept). With the builder standalone, migrating the factory only deletes the
delegate; the pybind, the Python fusion compiler, and mesh_partition keep calling
build_slice_tile_descriptor.

No behavior change. Verified: builds clean and the bind returns an identical
ProgramDescriptor (2 kernels, 1 CB) on device.
